riscv64,system,mmu: fix sign bug introduced in c3285ed6a

+++ src/isa/riscv64/system/mmu.c
@@ -79,10 +79,10 @@ static paddr_t ptw(vaddr_t vaddr, int type) {
   word_t p_pte; // pte pointer
   PTE pte;
   int level;
-  uint64_t vaddr39 = vaddr << (64 - 39);
+  int64_t vaddr39 = vaddr << (64 - 39);
   vaddr39 >>= (64 - 39);
   bool is_write = (type == MEM_TYPE_WRITE);
-  if (vaddr39 != vaddr) goto bad;
+  if ((uint64_t)vaddr39 != vaddr) goto bad;
   for (level = PTW_LEVEL - 1; level >= 0;) {
     p_pte = pg_base + VPNi(vaddr, level) * PTE_SIZE;
     pte.val	= paddr_read(p_pte, PTE_SIZE);
